+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.0.2 - 1.21.1
+- fix waterbomb passive lang formatting
+- fix Frost Resistance Status Effect
+- add Wither's Curse Status Effect
+
 ### 1.0.1 - 1.21.1
 - tweak some passive spells
 - fix -> GlazeEntity (ClassCastException with FreezingEffect on 1.21.1) #3